Fixes after merge
