+++ crates/solvers-dto/src/auction.rs
@@ -66,36 +66,42 @@ pub struct Order {
     #[schema(value_type = TokenAmount)]
     #[serde_as(as = "HexOrDecimalU256")]
     pub sell_amount: U256,
+    #[schema(value_type = TokenAmount)]
     #[serde_as(as = "HexOrDecimalU256")]
     pub full_sell_amount: U256,
     #[schema(value_type = TokenAmount)]
     #[serde_as(as = "HexOrDecimalU256")]
     pub buy_amount: U256,
+    #[schema(value_type = TokenAmount)]
     #[serde_as(as = "HexOrDecimalU256")]
     pub full_buy_amount: U256,
     pub fee_policies: Option<Vec<FeePolicy>>,
     pub valid_to: u32,
     pub kind: OrderKind,
+    #[schema(value_type = Address)]
     pub receiver: Option<H160>,
+    #[schema(value_type = Address)]
     pub owner: H160,
     /// Whether or not this order can be partially filled. If this is false,
     /// then the order is a "fill-or-kill" order, meaning it needs to be
     /// completely filled or not at all.
     pub partially_fillable: bool,
-     pub pre_interactions: Vec<InteractionData>,
+    pub pre_interactions: Vec<InteractionData>,
     pub post_interactions: Vec<InteractionData>,
     pub sell_token_source: SellTokenSource,
     pub buy_token_destination: BuyTokenDestination,
     pub class: OrderClass,
+    #[schema(value_type = AppData)]
     pub app_data: AppDataHash,
-    pub signing_scheme: SigningScheme,
+    pub signing_scheme: LegacySigningScheme,
     #[serde(with = "bytes_hex")]
+    #[schema(value_type = Signature)]
     pub signature: Vec<u8>,
 }
 
 /// Destination for which the buyAmount should be transferred to order's
 /// receiver to upon fulfillment
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, ToSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum BuyTokenDestination {
     /// Pay trade proceeds as an ERC20 token transfer
@@ -105,7 +111,7 @@ pub enum BuyTokenDestination {
 }
 
 /// Source from which the sellAmount should be drawn upon order fulfillment
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, ToSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum SellTokenSource {
     /// Direct ERC20 allowances to the Vault relayer contract
@@ -117,19 +123,25 @@ pub enum SellTokenSource {
 }
 
 #[serde_as]
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct InteractionData {
+    #[schema(value_type = Address)]
     pub target: H160,
+    #[schema(value_type = TokenAmount)]
     #[serde_as(as = "HexOrDecimalU256")]
     pub value: U256,
+    #[schema(value_type = String, example = "0x01020304")]
     #[serde(with = "bytes_hex")]
     pub call_data: Vec<u8>,
 }
 
-#[derive(Debug, Deserialize)]
+// todo: There is a conflict between solution's SigningScheme which is in
+// camelCase. There is no way to keep 2 object with the same name in the OpenAPI
+// schema. Temporarily renamed the struct. Must be migrated to the camelCase.
+#[derive(Debug, Deserialize, ToSchema)]
 #[serde(rename_all = "lowercase")]
-pub enum SigningScheme {
+pub enum LegacySigningScheme {
     Eip712,
     EthSign,
     Eip1271,
@@ -695,6 +707,7 @@ pub struct ForeignLimitOrder {
     #[serde_as(as = "HexOrDecimalU256")]
     pub gas_estimate: U256,
     #[serde_as(as = "serialize::Hex")]
+    // todo: not used/not a part of the API schema for some reason
     pub hash: [u8; 32],
     pub maker_token: H160,
     pub taker_token: H160,
@@ -746,23 +759,11 @@ impl ToSchema<'static> for ForeignLimitOrder {
 #[allow(dead_code)]
 pub struct NativePrice(String);
 
-/// Amount of an ERC20 token. 256 bit unsigned integer in decimal notation.
-#[derive(ToSchema)]
-#[schema(example = "1234567890")]
-#[allow(dead_code)]
-pub struct TokenAmount(String);
-
 /// An ISO-8601 formatted date-time.
 #[derive(ToSchema)]
 #[schema(example = "1970-01-01T00:00:00.000Z")]
 #[allow(dead_code)]
 pub struct DateTime(String);
-
-/// An Ethereum public address.
-#[derive(ToSchema)]
-#[schema(example = "0x0000000000000000000000000000000000000000")]
-#[allow(dead_code)]
-pub struct Address(String);
 
 /// An arbitrary-precision integer value.
 #[derive(ToSchema)]
@@ -782,12 +783,6 @@ pub struct Decimal(String);
 #[schema(example = "0xc88c76dd8b92408fe9bea1a54922a31e232d873c0002000000000000000005b2")]
 #[allow(dead_code)]
 pub struct BalancerPoolId(String);
-
-/// An ERC20 token address.
-#[derive(ToSchema)]
-#[schema(example = "0xDEf1CA1fb7FBcDC777520aa7f396b4E015F497aB")]
-#[allow(dead_code)]
-pub struct Token(String);
 
 #[serde_as]
 #[derive(ToSchema, Deserialize)]
@@ -831,6 +826,14 @@ pub struct I32(String);
 )]
 #[allow(dead_code)]
 pub struct OrderUid(String);
+
+/// Signature bytes.
+#[derive(ToSchema)]
+#[schema(
+    example = "0x0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+)]
+#[allow(dead_code)]
+pub struct Signature(String);
 
 /// If the order receives more than limit price, pay the protocol a factor of
 /// the difference.
