+++ crates/solvers/src/api/mod.rs
@@ -66,8 +66,6 @@ pub fn generate_openapi_yaml() -> Result<String, serde_yaml::Error> {
             solvers_dto::auction::ConcentratedLiquidityPool,
             solvers_dto::auction::ForeignLimitOrder,
             solvers_dto::auction::TokenReserve,
-            solvers_dto::auction::TokenAmount,
-            solvers_dto::auction::Token,
             solvers_dto::auction::BalancerPoolId,
             solvers_dto::auction::Decimal,
             solvers_dto::auction::U256Schema,
@@ -77,7 +75,6 @@ pub fn generate_openapi_yaml() -> Result<String, serde_yaml::Error> {
             solvers_dto::auction::BigInt,
             solvers_dto::auction::Order,
             solvers_dto::auction::OrderUid,
-            solvers_dto::auction::Address,
             solvers_dto::auction::FeePolicy,
             solvers_dto::auction::Quote,
             solvers_dto::auction::OrderClass,
@@ -85,6 +82,11 @@ pub fn generate_openapi_yaml() -> Result<String, serde_yaml::Error> {
             solvers_dto::auction::SurplusFee,
             solvers_dto::auction::PriceImprovement,
             solvers_dto::auction::VolumeFee,
+            solvers_dto::auction::SellTokenSource,
+            solvers_dto::auction::InteractionData,
+            solvers_dto::auction::BuyTokenDestination,
+            solvers_dto::auction::SellTokenSource,
+            solvers_dto::auction::LegacySigningScheme,
             solvers_dto::solution::Solution,
             solvers_dto::solution::Interaction,
             solvers_dto::solution::CustomInteraction,
@@ -95,11 +97,14 @@ pub fn generate_openapi_yaml() -> Result<String, serde_yaml::Error> {
             solvers_dto::solution::Fulfillment,
             solvers_dto::solution::JitTrade,
             solvers_dto::solution::JitOrder,
-            solvers_dto::solution::AppData,
             solvers_dto::solution::BuyTokenBalance,
             solvers_dto::solution::SellTokenBalance,
-            solvers_dto::solution::Signature,
             solvers_dto::solution::SigningScheme,
+            solvers_dto::common::Address,
+            solvers_dto::common::AppData,
+            solvers_dto::common::Signature,
+            solvers_dto::common::TokenAmount,
+            solvers_dto::common::Token,
         )),
         info(
             description = "The API implemented by solver engines interacting with the reference \
